+++ src/views/PersonalHistory/PersonalHistory.js
@@ -1,53 +1,34 @@
-import React, {
-  useState,
-  useCallback,
-  useEffect,
-  useRef,
-  useImperativeHandle,
-  forwardRef,
-} from 'react';
+import React, { useState, useCallback, useEffect, useRef } from 'react';
 import { useHistory } from 'react-router-dom';
-
-import {
-  Typography,
-  Button,
-  Card,
-  Grid,
-  FormControl,
-  FormControlLabel,
-  FormLabel,
-  RadioGroup,
-  Radio,
-  CircularProgress,
-  Checkbox,
-  FormGroup,
-} from '@material-ui/core';
-
-import { useFormik } from 'formik';
-
+import { Typography, Button, CircularProgress } from '@material-ui/core';
 import CustomBreadcrumbs from 'components/CustomBreadcrumbs';
 import PatientInfo from 'components/PatientInfo';
-
 import { useToast } from 'hooks/toast';
 import { usePatient } from 'context/PatientContext';
-
 import api from 'services/api';
-
 import useStyles from './styles';
-import CardInfo from './components/CardInfo';
+import { customBreadcrumbsLinks } from './statics';
+import PersonalHistoryForm from './components/PersonalHistoryForm';
 
 const PersonalHistory = () => {
   const classes = useStyles();
+
   const history = useHistory();
+
   const { addToast } = useToast();
+
   const { patient } = usePatient();
 
   const formRef = useRef(null);
+
   const [buttonDisabled, setButtonDisabled] = useState(true);
 
   const [loading, setLoading] = useState(false);
+
   const [patientHistory, setPatientHistory] = useState({});
+
   const [usoDrogas, setUsoDrogas] = useState([]);
+
   const [drogas, setDrogas] = useState([]);
 
   const handleInfos = useCallback(async () => {
@@ -63,12 +44,10 @@ const PersonalHistory = () => {
       const responseHistory = await api.get(
         `/pacientes/${patient.id}/historico`,
       );
+
       setPatientHistory(responseHistory.data);
     } catch (err) {
-      if (err.response.status === 404) {
-        return null;
-      }
-
+      if (err.response.status === 404) return null;
       addToast({
         type: 'error',
         message: 'Erro ao tentar carregar informações, tente novamente',
@@ -97,16 +76,7 @@ const PersonalHistory = () => {
   return (
     <div className={classes.root}>
       <div className={classes.header}>
-        <CustomBreadcrumbs
-          links={[
-            { label: 'Meus pacientes', route: '/meus-pacientes' },
-            { label: 'Categorias', route: '/categorias' },
-            {
-              label: 'História Pessoal',
-              route: '/categorias/historia-pessoal',
-            },
-          ]}
-        />
+        <CustomBreadcrumbs links={customBreadcrumbsLinks} />
       </div>
 
       <div>
@@ -132,7 +102,7 @@ const PersonalHistory = () => {
         {loading ? (
           <CircularProgress />
         ) : (
-          <Form
+          <PersonalHistoryForm
             drogas={drogas}
             onChange={bool => handleChange(bool)}
             patientHistory={patientHistory}
@@ -146,316 +116,3 @@ const PersonalHistory = () => {
 };
 
 export default PersonalHistory;
-
-const Form = forwardRef((props, ref) => {
-  const { patientHistory, usoDrogas, drogas, onChange } = props;
-  const classes = useStyles();
-  const history = useHistory();
-  const { addToast } = useToast();
-  const { patient } = usePatient();
-
-  const handleSubmit = async values => {
-    try {
-      const historico = {};
-      values.situacao_uso_drogas_id
-        ? (historico.situacao_uso_drogas_id = Number(
-          values.situacao_uso_drogas_id,
-        ))
-        : (historico.situacao_uso_drogas_id = undefined);
-      values.drogas
-        ? (historico.drogas = values.drogas)
-        : (historico.drogas = undefined);
-      values.tabagismo
-        ? values.tabagismo === 'true'
-          ? (historico.tabagismo = true)
-          : (historico.tabagismo = false)
-        : (historico.tabagismo = undefined);
-      values.etilismo
-        ? values.etilismo === 'true'
-          ? (historico.etilismo = true)
-          : (historico.etilismo = false)
-        : (historico.etilismo = undefined);
-      await api.post(`/pacientes/${patient.id}/historico`, historico);
-
-      addToast({
-        type: 'success',
-        message: 'Dados salvos com sucesso',
-      });
-      history.push('/categorias');
-    } catch (err) {
-      console.log(err);
-      addToast({
-        type: 'error',
-        message: 'Erro ao tentar registrar história pessoal, tente novamente',
-      });
-    }
-  };
-
-  const formik = useFormik({
-    initialValues: {
-      tabagismo:
-        patientHistory.tabagismo != null
-          ? patientHistory.tabagismo.toString()
-          : '',
-      situacao_uso_drogas_id:
-        patientHistory.situacao_uso_drogas_id != null
-          ? patientHistory.situacao_uso_drogas_id.toString()
-          : '',
-      drogas: patientHistory.drogas?.map(droga => droga.id),
-      etilismo:
-        patientHistory.etilismo != null
-          ? patientHistory.etilismo.toString()
-          : '',
-    },
-    onSubmit: handleSubmit,
-  });
-
-  useImperativeHandle(
-    ref,
-    () => {
-      return {
-        submit: formik.handleSubmit,
-      };
-    },
-    [formik.handleSubmit],
-  );
-
-  function shallowEqual(object1, object2) {
-    const keys1 = Object.keys(object1);
-    const keys2 = Object.keys(object2);
-
-    if (keys1.length !== keys2.length) {
-      return false;
-    }
-
-    for (let key of keys1) {
-      if (object1[key] !== object2[key]) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  useEffect(() => {
-    onChange(shallowEqual(formik.values, formik.initialValues));
-  }, [formik.values, formik.initialValues, onChange]);
-
-  return (
-    <form onSubmit={formik.handleSubmit}>
-      <div className={classes.formContainer}>
-        <Card className={classes.form}>
-          <Grid
-            container
-            spacing={2}
-          >
-            <Grid
-              item
-              sm={12}
-            >
-              <CardInfo
-                items={[
-                  {
-                    label: 'Fumante Diário:',
-                    description: '1 cigarro ao dia por no mínimo 1 mês',
-                  },
-                  {
-                    label: 'Fumante ocasional:',
-                    description:
-                      'menos de 1 cigarro por dia por no mínimo 1 mês',
-                  },
-                  {
-                    label: 'Ex-fumante:',
-                    description: 'parou de fumar há pelo menos 1 mês',
-                  },
-                  {
-                    label: 'Não fumante:',
-                    description: 'nunca fumaram ou fumam há menos de 1 mês',
-                  },
-                ]}
-                title="Classificação do tabagismo segundo OMS:"
-              />
-            </Grid>
-            <Grid
-              item
-              sm={12}
-            >
-              <FormControl
-                className={classes.formGroup}
-                component="fieldset"
-              >
-                <FormLabel>
-                  <Typography variant="h4">Tabagismo</Typography>
-                </FormLabel>
-
-                <RadioGroup
-                  name="tabagismo"
-                  value={formik.values.tabagismo}
-                >
-                  <FormControlLabel
-                    control={
-                      <Radio
-                        onChange={formik.handleChange}
-                        value="true"
-                      />
-                    }
-                    label="Sim"
-                  />
-
-                  <FormControlLabel
-                    control={
-                      <Radio
-                        onChange={formik.handleChange}
-                        value="false"
-                      />
-                    }
-                    label="Não"
-                  />
-                </RadioGroup>
-              </FormControl>
-            </Grid>
-
-            <Grid
-              item
-              sm={12}
-            >
-              <FormControl
-                className={classes.formGroup}
-                component="fieldset"
-              >
-                <FormLabel>
-                  <Typography variant="h4">
-                    Em relação ao uso de drogas ilícitas, em que opção você se
-                    enquadra?
-                  </Typography>
-                </FormLabel>
-
-                <RadioGroup
-                  name="situacao_uso_drogas_id"
-                  value={formik.values.situacao_uso_drogas_id}
-                >
-                  {usoDrogas.map(item => (
-                    <FormControlLabel
-                      control={
-                        <Radio
-                          onChange={formik.handleChange}
-                          value={String(item.id)}
-                        />
-                      }
-                      key={String(item.id)}
-                      label={item.descricao}
-                    />
-                  ))}
-                </RadioGroup>
-              </FormControl>
-            </Grid>
-
-            <Grid
-              item
-              sm={12}
-            >
-              <FormControl
-                className={classes.formGroup}
-                component="fieldset"
-              >
-                <FormLabel>
-                  <Typography variant="h4">
-                    Em caso de uso de drogas (atual ou ex-usuário), descrever
-                    quais drogas
-                  </Typography>
-                </FormLabel>
-
-                <FormGroup
-                  name="drogas"
-                  value={formik.values.drogas}
-                >
-                  {drogas.map(item => (
-                    <FormControlLabel
-                      control={
-                        <Checkbox
-                          checked={formik.initialValues.drogas?.includes(
-                            item.id,
-                          )}
-                          onChange={formik.handleChange}
-                          value={item.id}
-                        />
-                      }
-                      key={String(item.id)}
-                      label={item.descricao}
-                      name="drogas"
-                    />
-                  ))}
-                </FormGroup>
-              </FormControl>
-            </Grid>
-
-            <Grid
-              item
-              sm={12}
-            >
-              <CardInfo
-                items={[
-                  {
-                    label: 'Etilista:',
-                    description:
-                      'consumo de pelo menos 1 unidade (ver abaixo) de qualquer bebida alcoólica por dia no último ano',
-                  },
-                  {
-                    label: 'Ex-etilista:',
-                    description:
-                      'já consumiu bebida alcoólica, mas parou de consumir no último ano',
-                  },
-                  {
-                    label: 'Não etilista:',
-                    description:
-                      'nunca consumiu bebida alcoólica na frequência de etilista',
-                  },
-                ]}
-                title="Classificação do etilismo segundo OMS:"
-              />
-            </Grid>
-            <Grid
-              item
-              sm={12}
-            >
-              <FormControl
-                className={classes.formGroup}
-                component="fieldset"
-              >
-                <FormLabel>
-                  <Typography variant="h4">Etilismo</Typography>
-                </FormLabel>
-
-                <RadioGroup
-                  name="etilismo"
-                  value={formik.values.etilismo}
-                >
-                  <FormControlLabel
-                    control={
-                      <Radio
-                        onChange={formik.handleChange}
-                        value="true"
-                      />
-                    }
-                    label="Etilista / Ex- etilista"
-                  />
-                  <FormControlLabel
-                    control={
-                      <Radio
-                        onChange={formik.handleChange}
-                        value="false"
-                      />
-                    }
-                    label="Não etilista"
-                  />
-                </RadioGroup>
-              </FormControl>
-            </Grid>
-          </Grid>
-        </Card>
-      </div>
-    </form>
-  );
-});
-
-
