+++ src/views/PersonalHistory/statics.js
@@ -1,0 +1,63 @@
+export const cardInfoTabagismoItens = [
+  {
+    label: 'Etilista:',
+    description:
+      'consumo de pelo menos 1 unidade (ver abaixo) de qualquer bebida alcoólica por dia no último ano',
+  },
+  {
+    label: 'Ex-etilista:',
+    description:
+      'já consumiu bebida alcoólica, mas parou de consumir no último ano',
+  },
+  {
+    label: 'Não etilista:',
+    description: 'nunca consumiu bebida alcoólica na frequência de etilista',
+  },
+];
+
+export const cardInfoEtilismoItens = [
+  {
+    label: 'Etilista:',
+    description:
+      'consumo de pelo menos 1 unidade (ver abaixo) de qualquer bebida alcoólica por dia no último ano',
+  },
+  {
+    label: 'Ex-etilista:',
+    description:
+      'já consumiu bebida alcoólica, mas parou de consumir no último ano',
+  },
+  {
+    label: 'Não etilista:',
+    description: 'nunca consumiu bebida alcoólica na frequência de etilista',
+  },
+];
+
+export const customBreadcrumbsLinks = [
+  { label: 'Meus pacientes', route: '/meus-pacientes' },
+  { label: 'Categorias', route: '/categorias' },
+  {
+    label: 'História Pessoal',
+    route: '/categorias/historia-pessoal',
+  },
+];
+
+export const tabagismoOptions = [
+  { id: 1, name: 'Fumante' },
+  { id: 2, name: 'Ex-Fumante' },
+  { id: 3, name: 'Não Fumante' },
+  { id: 4, name: 'Não Informado' },
+];
+
+export const etilismoOptions = [
+  { id: 1, name: 'Etilista' },
+  { id: 2, name: 'Ex-etilista' },
+  { id: 3, name: 'Não Etilista' },
+  { id: 4, name: 'Não Informado' },
+];
+
+export const drogasOptions = [
+  { id: 1, name: 'É usuário' },
+  { id: 2, name: 'Ex-usuário' },
+  { id: 3, name: 'Nunca usou Drogas' },
+  { id: 4, name: 'Não Informado' },
+];
